requirements.txt
# Python Dependencies for Requirements Elicitation Pipeline

# Core LLM Integration
google-generativeai==0.8.6

# Data Processing
pydantic==2.12.5
pyyaml==6.0.1

# Environment Management
python-dotenv==1.2.1

# Utilities
numpy==2.4.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0

# Optional: For enhanced logging and debugging
colorlog==6.9.0
